Event handling, separate computer class
